refactor: SettlementByStoreRequest 서비스 폴더로 이동
