+++ src/main/java/com/ray/pominowner/settlement/controller/SettlementController.java
@@ -2,7 +2,7 @@ package com.ray.pominowner.settlement.controller;
 
 import com.ray.pominowner.settlement.controller.dto.DailySettlementResponse;
 import com.ray.pominowner.settlement.controller.dto.SettlementResponse;
-import com.ray.pominowner.settlement.controller.vo.SettlementByStoreRequest;
+import com.ray.pominowner.settlement.service.vo.SettlementByStoreRequest;
 import com.ray.pominowner.settlement.domain.Settlement;
 import com.ray.pominowner.settlement.service.SettlementService;
 import lombok.RequiredArgsConstructor;
