+++ src/main/java/com/ray/pominowner/settlement/service/DateType.java
@@ -1,6 +1,6 @@
 package com.ray.pominowner.settlement.service;
 
-import com.ray.pominowner.settlement.controller.vo.SettlementByStoreRequest;
+import com.ray.pominowner.settlement.service.vo.SettlementByStoreRequest;
 import com.ray.pominowner.settlement.domain.Settlement;
 import com.ray.pominowner.settlement.repository.SettlementRepository;
